refactor(qr-scanner): Migrate from QrReader to BarcodeScanner
The `react-zxing` library has deprecated the `QrReader` component in favor of the more versatile `BarcodeScanner` component.

This commit updates the `QrScanner` component to use the new `BarcodeScanner`. This involves changing the component name, updating the prop for handling scan results from `onResult` to `onUpdate`, and adjusting the structure of the `constraints` prop.

Additionally, the standalone route for the QR scanner (`/vib/qrscanner`) has been removed as the functionality is intended to be integrated within other pages rather than being a standalone feature.